+++ test/base_collection/dummy_collection.ts
@@ -12,4 +12,20 @@ export class DummyCollection extends BaseCollection<Branch> {
 	public testGetUri(uri: string | null): string {
 		return this.getUri(uri);
 	}
+
+	public testRootElementName(): string {
+		return this.rootElementName;
+	}
+
+	public testRootElementNameSingular(): string {
+		return this.rootElementNameSingular;
+	}
+
+	public testSecondaryElementClass(): new (json: Keyable) => Branch {
+		return this.secondaryElementClass;
+	}
+
+	public testSecondaryElementNameSingular(): string {
+		return this.secondaryElementNameSingular;
+	}
 }